Project/requirements.txt
google-generativeai
python-dotenv
IPython
cohere
requests
joblib
matplotlib 
pytesseract 
Pillow
scikit-learn
pandas
